+++ packages/adapter-statsig/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @flags-sdk/statsig
 
+## 0.2.1
+
+### Patch Changes
+
+- 9a687cb: accept consoleApiKey from getProviderData
+
+## 0.2.0
+
+### Minor Changes
+
+- bd7e10a: Initial support for Feature Gates and Dynamic Configs
+
 ## 0.1.0
 
 ### Minor Changes